+++ rskj-core/src/main/java/co/rsk/pcc/altBN128/AltBN128.java
@@ -1,176 +1,37 @@
 package co.rsk.pcc.altBN128;
 
-import co.rsk.altbn128.cloudflare.JniBn128;
 import co.rsk.altbn128.cloudflare.Utils;
-import co.rsk.crypto.altbn128java.*;
-import org.ethereum.util.BIUtil;
-import org.ethereum.vm.DataWord;
+import co.rsk.pcc.altBN128.impls.AbstractAltBN128;
+import co.rsk.pcc.altBN128.impls.GoAltBN128;
+import co.rsk.pcc.altBN128.impls.JavaAltBN128;
 
-import static co.rsk.pcc.altBN128.BN128Pairing.PAIR_SIZE;
-import static org.ethereum.util.ByteUtil.*;
+public class AltBN128 extends AbstractAltBN128 {
+    private AbstractAltBN128 implementation;
 
-public class AltBN128 {
-    private byte[] output;
-
-    private static byte[] encodeRes(byte[] w1, byte[] w2) {
-
-        byte[] res = new byte[64];
-
-        w1 = stripLeadingZeroes(w1);
-        w2 = stripLeadingZeroes(w2);
-
-        System.arraycopy(w1, 0, res, 32 - w1.length, w1.length);
-        System.arraycopy(w2, 0, res, 64 - w2.length, w2.length);
-
-        return res;
+    public AltBN128() {
+        if (Utils.isLinux()) {
+            this.implementation = new GoAltBN128();
+        } else {
+            this.implementation = new JavaAltBN128();
+        }
     }
-
+    @Override
     public int add(byte[] data, int length) {
-        if (Utils.isLinux()) {
-            byte[] jniOutput = new byte[64];
-            int res = new JniBn128().add(data, length, jniOutput);
-            output = jniOutput;
-            return res;
-        } else {
-            return addJava(data);
-        }
+        return implementation.add(data,length);
     }
 
+    @Override
     public int mul(byte[] data, int length) {
-        if (Utils.isLinux()) {
-            byte[] jniOutput = new byte[64];
-            int res = new JniBn128().mul(data, length, jniOutput);
-            output = jniOutput;
-            return res;
-        } else {
-            return mulJava(data);
-        }
+        return implementation.mul(data,length);
     }
 
+    @Override
     public int pairing(byte[] data, int length) {
-        if (Utils.isLinux()) {
-            byte[] jniOutput = new byte[32];
-            int res = new JniBn128().pairing(data, length, jniOutput);
-            output = jniOutput;
-            return res;
-        } else {
-            return pairingJava(data);
-        }
+        return implementation.pairing(data,length);
     }
 
-    public int addJava(byte[] data) {
-        output = new byte[64];
-
-        byte[] x1 = parseWord(data, 0);
-        byte[] y1 = parseWord(data, 1);
-
-        byte[] x2 = parseWord(data, 2);
-        byte[] y2 = parseWord(data, 3);
-
-        BN128<Fp> p1 = BN128Fp.create(x1, y1);
-
-        if (p1 == null) {
-            output = EMPTY_BYTE_ARRAY;
-            return 1;
-        }
-
-        BN128<Fp> p2 = BN128Fp.create(x2, y2);
-        if (p2 == null) {
-            output = EMPTY_BYTE_ARRAY;
-            return 1;
-        }
-
-        BN128<Fp> res = p1.add(p2).toEthNotation();
-
-        output = encodeRes(res.x().bytes(), res.y().bytes());
-        return 1;
-    }
-
-    public int mulJava(byte[] data) {
-        output = new byte[64];
-
-        byte[] x = parseWord(data, 0);
-        byte[] y = parseWord(data, 1);
-
-        byte[] s = parseWord(data, 2);
-
-        BN128<Fp> p = BN128Fp.create(x, y);
-
-        if (p == null) {
-            output = EMPTY_BYTE_ARRAY;
-            return 1;
-        }
-
-        BN128<Fp> res = p.mul(BIUtil.toBI(s)).toEthNotation();
-
-        output = encodeRes(res.x().bytes(), res.y().bytes());
-        return 1;
-    }
-
-    public int pairingJava(byte[] data) {
-
-        output = new byte[32];
-
-        // fail if input len is not a multiple of PAIR_SIZE
-        if (data.length % PAIR_SIZE > 0) {
-            output = EMPTY_BYTE_ARRAY;
-            return 1;
-        }
-
-        PairingCheck check = PairingCheck.create();
-
-        // iterating over all pairs
-        for (int offset = 0; offset < data.length; offset += PAIR_SIZE) {
-
-            BN128Pair pair = decodePair(data, offset);
-
-            // fail if decoding has failed
-            if (pair == null) {
-                output = EMPTY_BYTE_ARRAY;
-                return 1;
-            }
-
-            check.addPair(pair.getG1(), pair.getG2());
-        }
-
-        check.run();
-        int result = check.result();
-
-        output = DataWord.valueOf(result).getData();
-        return 1;
-    }
-
-    private BN128Pair decodePair(byte[] in, int offset) {
-
-        byte[] x = parseWord(in, offset, 0);
-        byte[] y = parseWord(in, offset, 1);
-
-        BN128G1 p1 = BN128G1.create(x, y);
-
-        // fail if point is invalid
-        if (p1 == null) {
-            return null;
-        }
-
-        // (b, a)
-        byte[] b = parseWord(in, offset, 2);
-        byte[] a = parseWord(in, offset, 3);
-
-        // (d, c)
-        byte[] d = parseWord(in, offset, 4);
-        byte[] c = parseWord(in, offset, 5);
-
-        BN128G2 p2 = BN128G2.create(a, b, c, d);
-
-        // fail if point is invalid
-        if (p2 == null) {
-            return null;
-        }
-
-        return BN128Pair.of(p1, p2);
-    }
-
+    @Override
     public byte[] getOutput() {
-        return output.clone();
+        return implementation.getOutput();
     }
 }
